+++ src/OpenTelemetry.Extensions/CHANGELOG.md
@@ -7,6 +7,13 @@
   [tracing API](https://github.com/open-telemetry/opentelemetry-specification/blob/v1.60.0/specification/trace/api.md#tracestate)
   specification defines, and preserving the valid members it did not generate.
   ([#5115](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5115))
+  
+## 1.19.0-beta.1
+
+Released 2026-Sep-18
+
+* Updated OpenTelemetry core component version(s) to `1.19.0`.
+  ([#5240](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/5240))
 
 ## 1.18.0-beta.1